fixed tests

+++ app/controllers/meal_arrangements_controller.rb
@@ -1,15 +1,7 @@
 class MealArrangementsController < ApplicationController
 
   before_filter :get_meal, :only => [:edit, :show, :update, :destroy]
-  before_filter :check_user, :only => [:edit, :update, :destroy]
-
-  def index
-    @meal_arrangements = MealArrangement.all
-  end
-
-  def new
-    @meal_arrangement = MealArrangement.new
-  end
+  before_filter :check_user, :only => [:update]
 
   def create
     @meal_arrangement = MealArrangement.new(:user_id => current_user.id,
@@ -23,22 +15,35 @@ class MealArrangementsController < ApplicationController
     end
   end
 
-
   def update
     if @meal_arrangement.update_attribute(:acceptance, true)
-        current_user.send_message(@meal_arrangement.user, I18n.t('message.accept'), I18n.t('message.info') )
-        redirect_to user_path(current_user.id), :notice => I18n.t('meal_arrangements.accept_success')
-      else
-        redirect_to user_path(current_user.id), :notice => I18n.t('meal_arrangements.accept_fail')
+      current_user.send_message(@meal_arrangement.user, I18n.t('message.accept'), I18n.t('message.info') )
+      redirect_to :back, :notice => I18n.t('meal_arrangements.accept_success')
+    else
+      redirect_to :back, :notice => I18n.t('meal_arrangements.accept_fail')
     end
   end
 
   def destroy
-    if @meal_arrangement.destroy
-      current_user.send_message(@meal_arrangement.user, I18n.t('message.reject'), I18n.t('message.sorry') )
-      redirect_to user_path(current_user.id), :notice => I18n.t('meal_arrangements.delete_success') 
+    if current_user.id == @meal_arrangement.meal.user_id #Cook
+      if @meal_arrangement.destroy
+        current_user.send_message(@meal_arrangement.user, I18n.t('message.reject'), I18n.t('message.sorry') )
+        redirect_to :back, :notice => I18n.t('meal_arrangements.delete_success')
+      else
+        redirect_to :back, :notice => I18n.t('meal_arrangements.delete_fail')
+      end
+    elsif current_user.id == @meal_arrangement.user_id #Eater
+      unless @meal_arrangement.acceptance
+        if @meal_arrangement.destroy
+          redirect_to :back, :notice => I18n.t('meal_arrangements.delete_success')
+        else
+          redirect_to :back, :notice => I18n.t('meal_arrangements.delete_fail')
+        end
+      else
+        redirect_to :back, :notice => I18n.t('meal_arrangements.acceptance_given_fail')
+      end
     else
-        redirect_to user_path(current_user.id), :notice => I18n.t('meal_arrangements.delete_fail') 
+      redirect_to root_path, :notice => I18n.t('application.rights_fail')
     end
   end
